requirements.txt
accelerate==1.12.0
faiss-cpu==1.13.2
fastapi==0.129.0
pypdf==6.7.0
python-docx==1.2.0
sentence-transformers==5.2.2
torch==2.10.0
transformers==5.1.0
uvicorn==0.40.0
pydantic-settings==2.12.0
python-multipart==0.0.22
